# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Buccaneer Wallet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/balance:
    get:
      operationId: getBalance
      tags: [wallet]
      summary: Get wallet balance
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /wallet/send:
    post:
      operationId: sendPayment
      tags: [wallet]
      summary: Send Lightning payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendPaymentRequest"
      responses:
        "200":
          description: Payment sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendPaymentResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /wallet/decode-invoice:
    post:
      operationId: decodeInvoice
      tags: [wallet]
      summary: Decode a Lightning invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecodeInvoiceRequest"
      responses:
        "200":
          description: Decoded invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecodedInvoice"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /wallet/receive:
    post:
      operationId: createInvoice
      tags: [wallet]
      summary: Create Lightning invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "200":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateInvoiceResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /wallet/btc-price:
    get:
      operationId: getBtcPrice
      tags: [wallet]
      summary: Get current BTC price in fiat
      parameters:
        - name: currency
          in: query
          schema:
            type: string
            default: USD
      responses:
        "200":
          description: BTC price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BtcPrice"
...
components:
  securitySchemes:
    ownerSessionHeader:
      type: apiKey
      in: header
      name: X-Owner-Session
    ownerBootstrapHeader:
      type: apiKey
      in: header
      name: X-Wallet-Owner
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    WalletBalance:
      type: object
      properties:
        balanceSats:
          type: integer
          description: Balance in satoshis
        pendingSendSats:
          type: integer
        pendingReceiveSats:
          type: integer
      required: [balanceSats, pendingSendSats, pendingReceiveSats]
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [send, receive]
        amountSats:
          type: integer
        feeSats:
          type: integer
        description:
          type: string
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, complete, failed]
        paymentHash:
          type: string
      required: [id, type, amountSats, feeSats, timestamp, status]
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
      required: [transactions, total]
    SendPaymentRequest:
      type: object
      properties:
        bolt11:
          type: string
          description: BOLT11 Lightning invoice
        amountSats:
          type: integer
          description: Amount in sats (for zero-amount invoices)
      required: [bolt11]
    SendPaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        paymentHash:
          type: string
        feeSats:
          type: integer
        amountSats:
          type: integer
      required: [success, paymentHash, feeSats, amountSats]
    DecodeInvoiceRequest:
      type: object
      properties:
        bolt11:
          type: string
      required: [bolt11]
    DecodedInvoice:
      type: object
      properties:
        amountSats:
          type: integer
        description:
          type: string
        expiry:
          type: integer
        paymentHash:
          type: string
        isExpired:
          type: boolean
      required: [isExpired]
    CreateInvoiceRequest:
      type: object
      properties:
        amountSats:
          type: integer
        description:
          type: string
      required: [amountSats]
    CreateInvoiceResponse:
      type: object
      properties:
        bolt11:
          type: string
        paymentHash:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [bolt11, paymentHash]
...
    BtcPrice:
      type: object
      properties:
        currency:
          type: string
        price:
          type: number
        symbol:
          type: string
      required: [currency, price, symbol]
...